requirements.txt
annotated-types==0.7.0
anyio==4.4.0
exceptiongroup==1.2.0
fastapi==0.115.0
fake-useragent==2.1.0
idna==3.8
numpy==1.26.4
Pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
sniffio==1.3.1
starlette==0.38.6
torch==2.1.0
typing_extensions==4.12.2
transformers==4.32.1
uvicorn==0.30.6
python-multipart==0.0.20
pillow-avif-plugin==1.5.2
cairosvg==2.7.1
python-dotenv==1.0.1